backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
google-api-python-client
pydantic
